pyproject.toml
[tool.poetry]
name = "sleeping-agent-game-demo"
version = "0.1.0"
description = ""
authors = ["Kirill Yakunin <yakunin.k@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.41.1"
openai = "^1.59.7"
python-dotenv = "^1.0.1"
anthropic = "^0.45.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
